Color picker improvements + color picker for all of text color
FEAT: Added the ability to pick a color for a text block's entire color(the hex code at the top). Pressing cancel will revert the color back to before the color picker was moved, and pressing accept will close the picker and unfocus the text input.
REFACTOR: Updated how accepting, canceling, and choosing a preset is handled with the color picker, allowing for different code to be executed at different times